requirements.txt
# Runtime
pygame-ce>=2.5,<3.0
chess>=1.11,<2.0
anthropic>=0.67,<1.0
python-dotenv>=1.0,<2.0

# Tests
pytest>=8.0,<9.0
